requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
appnope==0.1.3
asttokens==2.2.1
asyncpg==0.27.0
backcall==0.2.0
black==22.8.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.6
ecdsa==0.18.0
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.95.1
filelock==3.12.0
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
ipykernel==6.23.0
ipython==8.13.2
iso8601==1.1.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
nest-asyncio==1.5.6
numpy==1.24.2
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pydantic==1.10.7
Pygments==2.15.1
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-jose==3.3.0
python-keycloak-client==0.2.3
pytz==2022.7.1
pyzmq==25.0.2
requests==2.28.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.7
stack-data==0.6.2
starlette==0.26.1
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.3.1
tortoise-orm==0.19.3
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
virtualenv==20.23.0
wcwidth==0.2.6
Werkzeug==2.2.3
